chore: sync FrotiX.Site.OLD - Models, Controllers, Repository, Services e Scripts SQL
- Controllers: AlertasFrotiXController, TaxiLegController
- Models: AlertasFrotiX, Abastecimento, AtaRegistroPrecos, Contrato, CorridasTaxiLeg, Lavagem, Motorista, Viagem, VeiculoPadraoViagem, ViagemEstatistica
- Pages: AlertasFrotiX/Upsert.cshtml.cs
- Repository: AlertasFrotiXRepository, VeiculoPadraoViagemRepository
- Services: AlertasBackgroundService
- Scripts_SQL: Limpeza_OrigemDestino_COMPLETO.sql

+++ FrotiX.Site.OLD/Repository/AlertasFrotiXRepository.cs
@@ -266,16 +266,16 @@ namespace FrotiX.Repository
                     .Where(a => a.Ativo &&
                            (
                                // Alertas que devem ser exibidos ao abrir o sistema
-                               (a.TipoExibicao == TipoExibicaoAlerta.AoAbrir) ||
+                               (a.TipoExibicao == (int)TipoExibicaoAlerta.AoAbrir) ||
 
                                // Alertas com horário específico (verifica se chegou a hora hoje)
-                               (a.TipoExibicao == TipoExibicaoAlerta.Horario &&
+                               (a.TipoExibicao == (int)TipoExibicaoAlerta.Horario &&
                                 a.HorarioExibicao.HasValue &&
                                 agora.TimeOfDay >= a.HorarioExibicao.Value &&
                                 agora.TimeOfDay <= a.HorarioExibicao.Value.Add(TimeSpan.FromMinutes(5))) ||
 
                                // Alertas com data/hora específica
-                               (a.TipoExibicao == TipoExibicaoAlerta.DataHora &&
+                               (a.TipoExibicao == (int)TipoExibicaoAlerta.DataHora &&
                                 a.DataExibicao.HasValue &&
                                 a.DataExibicao.Value <= agora &&
                                 (!a.DataExpiracao.HasValue || a.DataExpiracao.Value >= agora))
